+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add HAProxy Ingress Controller support.
 - RealmManagement and Pairing have `get` and `list` permissions on
   endpoints, pods and services.
+- Add experimental support for FIDO Device Onboard (FDO) in Astarte Pairing. The feature
+  can be enabled and configured through the `features.fdo` field in the Astarte CR.
 
 ### Changed
 - Forward port changes from release-24.5